奖金计算

+++ armor-api/src/main/java/com/zhixinhuixue/armor/service/impl/ZSYUserWeekSevice.java
@@ -22,7 +22,7 @@ public class ZSYUserWeekSevice implements IZSYUserWeekService{
      * @return
      */
     @Override
-    public int getUserWeekHours(Long id, int weekNumber, int year) {
+    public Double getUserWeekHours(Long id, int weekNumber, int year) {
         return userWeekMapper.getUserWeekHours(id, weekNumber, year);
     }
 }
